Se creo el boton de eliminar producto y se actualiza el carrito sin recargar pagina.

+++ app.js
@@ -56,9 +56,18 @@ app.get('/productos', async(req, res) => {
   
 })
 
-
+//ruta para crear productos// 
 app.post('/productos', async(req, res) => {
   try {
+    //validar si el rol del usuario es ADMINISTRADOR//
+    if(!req.session.usuario){
+      res.status(401).json({error:"Inicia sesion primero"})
+      return
+    }
+    if(req.session.usuario.rol!="ADMINISTRADOR"){
+      res.status(401).json({error:"No tienes permisos para realizar esta accion"})
+      return
+    }
     console.log(req.body)
     const [results, fields] = await connection.query(
        "INSERT INTO `productos` (`id`, `nombre`, `descripcion`, `precio`, `imagen`) VALUES (NULL, ?, ?, ?, ?);",[req.body.nombre,req.body.descripcion,req.body.precio,req.body.imagen ]
@@ -75,6 +84,15 @@ app.post('/productos', async(req, res) => {
 //ruta para eliminar productos 
 app.delete('/productos/:id', async(req, res) => {
   try {
+    //validar si el rol del usuario es ADMINISTRADOR//
+    if(!req.session.usuario){
+      res.status(401).json({error:"Inicia sesion primero"})
+      return
+    }
+    if(req.session.usuario.rol!="ADMINISTRADOR"){
+      res.status(401).json({error:"No tienes permisos para realizar esta accion"})
+      return
+    }
     console.log(req.params.id)
     const [results, fields] = await connection.query(
        "DELETE FROM `productos` WHERE `productos`.`id` = ?;",[req.params.id]
@@ -163,6 +181,23 @@ app.get('/carrito', async(req, res) => {
   
 })
 
+//crear ruta para eliminar un producto del carrito//
+app.delete('/carrito/:id', async(req, res) => {
+  try {
+    console.log(req.params.id)
+    const [results, fields] = await connection.query(
+       "DELETE FROM `carrito` WHERE `carrito`.`id` = ?;",[req.params.id]
+    );
+    res.json(results)
+    // res.redirect('http://
+    console.log(results); // results contains rows returned by server
+    console.log(fields); // fields contains extra meta data about results, if available
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({error:"algo salio mal"})
+  }
+
+})
 
 app.listen(port, () => {
   console.log(`Example app listening on port ${port}`)
